Backend para finanzas creado y con avances, frontend para finanzas creado. Correcciones menores.

+++ front-end-new/src/app/shared/navbar/navbar.component.ts
@@ -36,7 +36,7 @@ export class NavbarComponent implements OnInit {
     private confirmationService: ConfirmationService,
     private messageService: MessageService,
     private authService: AuthService,
-    private usuarioService: UsuarioService,
+    private usuarioService: UsuarioService
   ) {}
 
   ngOnInit(): void {
@@ -46,7 +46,7 @@ export class NavbarComponent implements OnInit {
       !this.router.url.includes('login')
     ) {
       this.usuarioService
-        .getOwnUsuarioById(+this.cookieService.get('usuario_id'))
+        .getMinUsuarioById(+this.cookieService.get('usuario_id'))
         .subscribe({
           next: (res) => {
             this.usuario = res.result;
@@ -72,12 +72,18 @@ export class NavbarComponent implements OnInit {
                 'Ha ocurrido un error inesperado. Por favor, inicie sesión nuevamente.',
               life: 3000,
             });
-            this.router.navigate(['/login'], {queryParams: { redirect: true}});
-            this.cookieService.delete('usuario_act');
-            this.cookieService.delete('usuario_rol');
-            if (this.cookieService.get('usuario_id') === null) {
-              this.cookieService.delete('usuario_id');
-            }
+            this.router.navigate(['/login'], {
+              queryParams: { redirect: true },
+            }).then(() => {
+              this.cookieService.delete('usuario_act');
+              this.cookieService.delete('usuario_rol');
+              if (
+                this.cookieService.check('usuario_id') &&
+                this.cookieService.get('usuario_id') === null
+              ) {
+                this.cookieService.delete('usuario_id');
+              }
+            });
           },
         });
     }
@@ -97,7 +103,7 @@ export class NavbarComponent implements OnInit {
   buscarProductos() {
     if (this.searchForm.valid) {
       let b = this.searchForm.value;
-      this.router.navigate(['search', b.busqueda]);
+      this.redirectTo('search', b.busqueda);
     }
   }
 
@@ -128,5 +134,11 @@ export class NavbarComponent implements OnInit {
       },
       reject: () => {},
     });
+  }
+
+  redirectTo(uri: string, params: any) {
+    this.router
+      .navigateByUrl('/', { skipLocationChange: true })
+      .then(() => this.router.navigate([uri, params]));
   }
 }
